requirements.txt
# Spotter - Modular Reconnaissance Framework
# Core dependencies

# Network and DNS
dnspython==2.4.2
python-whois==0.8.0
requests==2.31.0

# Terminal output
colorama==0.4.6

# Optional but recommended
# nmap (system package - install via: apt-get install nmap)
